Updated the css styling for filter locations on location page
Updated the css styling for the filter on the locations page to be on one line with some more spacing now that the filter label as been added to the filter dropdown menu. The styling also puts the dropdown inline on other browsers rather than just safari which the intital code was tested on (to do so meant creating a fresh class for the container-inline). Also the filter created by html has had the label adjusted to go to one line rather than be on two lines